Move cohortextractor utils into their own module
Having it in the same module as the other utilities was causing problems when running tests due to having to import cohortextractor
