+++ analysis/utilities.py
@@ -4,8 +4,6 @@ import matplotlib.dates as mdates
 from matplotlib.lines import Line2D
 import numpy as np
 import pandas as pd
-from cohortextractor import patients, codelist
-
 from pathlib import Path
 
 BASE_DIR = Path(__file__).parents[1]
@@ -177,23 +175,4 @@ def relabel_sex(df):
     df = df.replace({"sex": sex_codes})
     return df
 
-def loop_over_codes(code_list):
-    def make_variable(code):
-        return {
-            f"count_{code}": (
-                patients.with_these_clinical_events(
-                    codelist([code], system="snomed"),
-                    between =["first_day_of_month(index_date)", "last_day_of_month(index_date)"],
-                    returning="number_of_matches_in_period",
-                    return_expectations={
-                         "incidence": 0.1,
-                         "int": {"distribution": "normal", "mean": 3, "stddev": 1},
-                    },
-                )
-            )
-        }
 
-    variables = {}
-    for code in code_list:
-        variables.update(make_variable(code))
-    return variables
